[refactor/#59]: 파일명 명시적으로 수정
