+++ types/studyTypes.ts
